+++ src/main/java/com/nifou/m/ifou_mobile_was/entity/result/Sub05Entity.java
@@ -30,7 +30,7 @@ public class Sub05Entity {
     private int totsales;
     @Column
     @JsonProperty("amt")
-    private int totreceivable;
+    private Long totreceivable;
 
     public Sub05Entity() {
