+++ src/main/java/com/nifou/m/ifou_mobile_was/entity/result/Sub06Entity.java
@@ -30,7 +30,7 @@ public class Sub06Entity {
     private String acqnm;
     @Column
     @JsonProperty("amt")
-    private int totamt;
+    private Long totamt;
 
     public Sub06Entity() {
